Move and rename DetectVendor, Numa, and MemoryResourceTraits
* Numa -> numa
* MemoryResourceTraits to util
* DetectVendor -> detect_vendor to util
