+++ src/umpire/ResourceManager.hpp
@@ -27,7 +27,6 @@
 #include "umpire/util/AllocationMap.hpp"
 
 #include "umpire/resource/MemoryResourceTypes.hpp"
-#include "umpire/resource/MemoryResourceTraits.hpp"
 
 namespace umpire {
