+++ src/umpire/util/detect_vendor.cpp
@@ -12,17 +12,19 @@
 // For details, see https://github.com/LLNL/Umpire
 // Please also see the LICENSE file for MIT license.
 //////////////////////////////////////////////////////////////////////////////
-#ifndef UMPIRE_DetectVendor_HPP
-#define UMPIRE_DetectVendor_HPP
 
-#include "umpire/resource/MemoryResourceTraits.hpp"
+#include "umpire/util/detect_vendor.hpp"
 
 namespace umpire {
-namespace resource {
 
-resource::MemoryResourceTraits::vendor_type cpu_vendor_type() noexcept;
+MemoryResourceTraits::vendor_type cpu_vendor_type() noexcept {
+#if defined(__x86_64__)
+  return MemoryResourceTraits::vendor_type::INTEL;
+#elif defined(__powerpc__)
+  return MemoryResourceTraits::vendor_type::IBM;
+#else
+  return MemoryResourceTraits::vendor_type::UNKNOWN;
+#endif
+}
 
-} // end namespace resource
 } // end namespace umpire
-
-#endif // UMPIRE_DetectVendor_HPP
